make animation of the work sectiion
