+++ src/components/Home/home.scss
@@ -1,38 +1,23 @@
 
 .section-main {
+    animation: fadeInUp 2s 2s;
     display: flex;
     flex-wrap: wrap;
     background-size: 150%;
-    height: 100vh;
-    animation: fadeInUp 2s 2s;
     animation-fill-mode: forwards;
     clear: both;
-
+    
     .side-by-side {
         margin: auto;
         width: 22%;
         height: 40%;
       }
-      h1 {
-        
-    }
     .full-side {
         flex: 1 0 100%;
-        animation: fadeInUp 2.5s 2s;
-        animation-fill-mode: forwards;
-        clear: both;
-
         &:hover {
-            animation: pulse 1s;
-            color: lightyellow
+            text-shadow: 0px 0px 5px lightslategrey;
         }
     }
-
-    // .stage-cube-cont {
-    //     .cubespinner {
-        
-    //     }
-    // }
 
     .cubespinner {
     animation-name: spincube;
@@ -99,3 +84,23 @@
 
     }
 }
+
+@keyframes pulse {
+    0% { transform: scale(1); }
+    50% { transform: scale(1.1); }
+    100% { transform: scale(1); }
+  }
+
+  .h-animate {
+    animation: none;
+    color: white;
+    transition: color 0.5s ease-in-out;
+}
+
+  .animate_pulse {
+    display: inline-block;
+    animation: pulse 1s infinite;
+    text-shadow: 0px 0px 10px lightslategrey;
+    transition: color 0.5s ease-in-out;
+}
+
