+++ src/components/Technologies/technologies.scss
@@ -1,0 +1,78 @@
+.tech-section {
+    animation: fadeInUp 2s 2s;
+    animation-fill-mode: forwards;
+    clear: both;
+    background:
+    url(https://res.cloudinary.com/dhko6cdgf/image/upload/v1704552054/josephIo/Images/Decoration/Desktop/rorazu3dmq1c1qnzrlwa.png) 0% 100% no-repeat,
+              url(https://res.cloudinary.com/dhko6cdgf/image/upload/v1704552052/josephIo/Images/Decoration/Desktop/rszgf9fmxizuj2pahcrc.svg) 100% 0% no-repeat;
+    background-size: 20%;
+
+.filter-white {
+    filter: invert(10%) sepia(100%) saturate(1%) hue-rotate(273deg) brightness(100%) contrast(1500%)
+}
+
+.filter-tailwind {
+    filter: brightness(120%) drop-shadow(10px 10px 5px #29f2f2);;
+}
+
+.filter-bootstrap {
+    filter: invert(14%) sepia(37%) saturate(7116%) hue-rotate(274deg) brightness(90%) contrast(114%) drop-shadow(3px 3px 20px #690899);
+}
+
+.filter-html {
+    filter: hue-rotate(30deg) saturate(200%) brightness(100%) drop-shadow(3px 3px 15px #eb9d02);
+}
+
+.filter-green {
+    filter: invert(48%) sepia(79%) saturate(2476%) hue-rotate(86deg) brightness(118%) contrast(119%);
+}
+
+.filter-css {
+    filter: invert(14%) sepia(56%) saturate(5640%) hue-rotate(219deg) brightness(94%) contrast(99%) drop-shadow(3px 3px 15px #0241b8);;
+}
+
+.filter-js {
+    filter: invert(78%) sepia(19%) saturate(2187%) hue-rotate(12deg) brightness(107%) drop-shadow(3px 3px 15px #e0dd07);;
+}
+
+.filter-ruby {
+    filter: invert(18%) sepia(68%) saturate(6074%) hue-rotate(15deg) brightness(101%) contrast(94%) drop-shadow(3px 3px 15px #e02f07);;
+}
+
+.filter-php {
+    filter: invert(30%) sepia(10%) saturate(2897%) hue-rotate(196deg) brightness(90%) contrast(80%) drop-shadow(3px 3px 15px #778ec9);;
+}
+
+.filter-rails {
+    filter: invert(15%) sepia(93%) saturate(7176%) hue-rotate(5deg) brightness(103%) contrast(118%) drop-shadow(3px 3px 15px #ff0000);
+}
+
+.filter-sql {
+    filter: invert(90%) sepia(13%) saturate(1796%) hue-rotate(352deg) brightness(98%) contrast(91%) drop-shadow(3px 3px 15px #f0d362);;
+}
+
+.filter-sass {
+    filter: invert(24%) sepia(46%) saturate(3892%) hue-rotate(289deg) brightness(81%) contrast(88%) drop-shadow(3px 3px 15px #c762f0);
+}
+
+.filter-react {
+    filter: invert(18%) sepia(100%) saturate(7422%) hue-rotate(242deg) brightness(80%) contrast(137%) drop-shadow(3px 3px 15px #47DEBD);
+}
+
+.filter-redux {
+    filter: invert(22%) sepia(96%) saturate(2411%) hue-rotate(253deg) brightness(96%) contrast(81%) drop-shadow(3px 3px 15px #7645d9);
+}
+
+
+.filter-mysql {
+    filter: invert(100%) sepia(100%) saturate(100%) hue-rotate(0deg) brightness(100%) contrast(100%) drop-shadow(3px 3px 15px #7645d9);;
+}
+
+.filter-postgres {
+    filter: invert(100%) sepia(100%) saturate(100%) hue-rotate(0deg) brightness(100%) contrast(100%) drop-shadow(3px 3px 15px #7645d9);;
+}
+
+.filter-jest {
+    filter: invert(9%) sepia(58%) saturate(4169%) hue-rotate(337deg) brightness(77%) contrast(98%) drop-shadow(3px 3px 15px #570D18);; 
+}
+}
